Extracted super class for common code
