+++ sql-plugin/src/main/330+/scala/org/apache/spark/sql/execution/datasources/parquet/ShimCurrentBatchIterator.scala
@@ -21,25 +21,20 @@ import java.lang.reflect.Method
 
 import scala.collection.JavaConverters._
 
-import com.nvidia.spark.rapids.{ByteArrayInputFile, ParquetCachedBatch}
+import com.nvidia.spark.CurrentBatchIterator
+import com.nvidia.spark.rapids.ParquetCachedBatch
 import java.util
 import org.apache.hadoop.conf.Configuration
-import org.apache.parquet.{ParquetReadOptions, VersionParser}
-import org.apache.parquet.VersionParser.ParsedVersion
-import org.apache.parquet.column.ColumnDescriptor
+import org.apache.parquet.ParquetReadOptions
 import org.apache.parquet.column.page.PageReadStore
-import org.apache.parquet.hadoop.ParquetFileReader
 import org.apache.parquet.schema.{GroupType, Type}
 
-import org.apache.spark.TaskContext
 import org.apache.spark.memory.MemoryMode
 import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.execution.vectorized.{OffHeapColumnVector, WritableColumnVector}
+import org.apache.spark.sql.execution.vectorized.WritableColumnVector
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.SQLConf.LegacyBehaviorPolicy
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.vectorized
-import org.apache.spark.sql.vectorized.ColumnarBatch
 
 object ParquetVectorizedReader {
   /**
@@ -65,94 +60,40 @@ object ParquetVectorizedReader {
  * org.apache/spark/sql/execution/datasources/parquet/VectorizedParquetRecordReader.java
  */
 class ShimCurrentBatchIterator(
-    val parquetCachedBatch: ParquetCachedBatch,
+    parquetCachedBatch: ParquetCachedBatch,
     conf: SQLConf,
     selectedAttributes: Seq[Attribute],
     options: ParquetReadOptions,
-    hadoopConf: Configuration) extends Iterator[ColumnarBatch] with AutoCloseable {
+    hadoopConf: Configuration)
+    extends CurrentBatchIterator(
+      parquetCachedBatch,
+      conf,
+      selectedAttributes,
+      options,
+      hadoopConf) {
 
-  var reqParquetSchemaInCacheOrder: org.apache.parquet.schema.MessageType = null
-  var inMemCacheParquetSchema: org.apache.parquet.schema.MessageType = null
-  val capacity = conf.parquetVectorizedReaderBatchSize
-  var parquetColumnVectors: Array[ParquetColumnVector] = _
-  var vectors: Array[OffHeapColumnVector] =
-    OffHeapColumnVector.allocateColumns(capacity, selectedAttributes.toStructType)
-  val columnarBatch = new ColumnarBatch(vectors.asInstanceOf[Array[vectorized.ColumnVector]])
-  var rowsReturned: Long = 0L
-  var numBatched = 0
-  var batchIdx = 0
-  var totalCountLoadedSoFar: Long = 0
-  val parquetFileReader = {
-    ParquetFileReader.open(new ByteArrayInputFile(parquetCachedBatch.buffer), options)
+  // reset spark schema calculated from parquet schema
+  hadoopConf.set(ParquetReadSupport.SPARK_ROW_REQUESTED_SCHEMA, inMemReqSparkSchema.json)
+  hadoopConf.set(ParquetWriteSupport.SPARK_ROW_SCHEMA, inMemReqSparkSchema.json)
+  val missingColumns: util.Set[ParquetColumn] = new util.HashSet[ParquetColumn]()
+  val config = new Configuration
+  config.setBoolean(SQLConf.PARQUET_BINARY_AS_STRING.key, false)
+  config.setBoolean(SQLConf.PARQUET_INT96_AS_TIMESTAMP.key, false)
+  config.setBoolean(SQLConf.CASE_SENSITIVE.key, false)
+  val parquetColumn = new ParquetToSparkSchemaConverter(config)
+      .convertParquetColumn(reqParquetSchemaInCacheOrder, Option.empty)
+
+  // initialize missingColumns to cover the case where requested column isn't present in the
+  // cache, which should never happen but just in case it does
+  for (column <- parquetColumn.children) {
+    checkColumn(column)
   }
-  val writerVersion: ParsedVersion = try {
-    VersionParser.parse(parquetFileReader.getFileMetaData.getCreatedBy)
-  } catch {
-    case _: Exception =>
-      // If any problems occur trying to parse the writer version, fallback to sequential reads
-      // if the column is a delta byte array encoding (due to PARQUET-246).
-      null
-  }
-  val (totalRowCount, columnsRequested, cacheSchemaToReqSchemaMap, missingColumns,
-  columnsInCache, typesInCache) = {
-    val parquetToSparkSchemaConverter = new ParquetToSparkSchemaConverter(hadoopConf)
-    // we are getting parquet schema and then converting it to catalyst schema
-    // because catalyst schema that we get from Spark doesn't have the exact schema expected
-    // by the columnar parquet reader
-    inMemCacheParquetSchema = parquetFileReader.getFooter.getFileMetaData.getSchema
-    val inMemCacheSparkSchema = parquetToSparkSchemaConverter.convert(inMemCacheParquetSchema)
 
-    val totalRowCount = parquetFileReader.getRowGroups.asScala.map(_.getRowCount).sum
-    val inMemReqSparkSchema = StructType(selectedAttributes.toStructType.map { field =>
-      inMemCacheSparkSchema.fields(inMemCacheSparkSchema.fieldIndex(field.name))
-    })
-    val reqSparkSchemaInCacheOrder = StructType(inMemCacheSparkSchema.filter(f =>
-      inMemReqSparkSchema.fields.exists(f0 => f0.name.equals(f.name))))
-
-    // There could be a case especially in a distributed environment where the requestedSchema
-    // and cacheSchema are not in the same order. We need to create a map so we can guarantee
-    // that we writing to the correct columnVector
-    val cacheSchemaToReqSchemaMap: Map[Int, Int] =
-    reqSparkSchemaInCacheOrder.indices.map { index =>
-      index -> inMemReqSparkSchema.fields.indexOf(reqSparkSchemaInCacheOrder.fields(index))
-    }.toMap
-
-    reqParquetSchemaInCacheOrder = new org.apache.parquet.schema.MessageType(
-      inMemCacheParquetSchema.getName(), reqSparkSchemaInCacheOrder.fields.map { f =>
-        inMemCacheParquetSchema.getFields().get(inMemCacheParquetSchema.getFieldIndex(f.name))
-      }:_*)
-
-    val columnsRequested: util.List[ColumnDescriptor] = reqParquetSchemaInCacheOrder.getColumns
-    // reset spark schema calculated from parquet schema
-    hadoopConf.set(ParquetReadSupport.SPARK_ROW_REQUESTED_SCHEMA, inMemReqSparkSchema.json)
-    hadoopConf.set(ParquetWriteSupport.SPARK_ROW_SCHEMA, inMemReqSparkSchema.json)
-
-    val columnsInCache: util.List[ColumnDescriptor] = reqParquetSchemaInCacheOrder.getColumns
-    val typesInCache: util.List[Type] = reqParquetSchemaInCacheOrder.asGroupType.getFields
-    val missingColumns: util.Set[ParquetColumn] = new util.HashSet[ParquetColumn]()
-
-    val config = new Configuration
-    config.setBoolean(SQLConf.PARQUET_BINARY_AS_STRING.key, false)
-    config.setBoolean(SQLConf.PARQUET_INT96_AS_TIMESTAMP.key, false)
-    config.setBoolean(SQLConf.CASE_SENSITIVE.key, false)
-    val parquetColumn = new ParquetToSparkSchemaConverter(config)
-        .convertParquetColumn(reqParquetSchemaInCacheOrder, Option.empty)
-
-    // initialize missingColumns to cover the case where requested column isn't present in the
-    // cache, which should never happen but just in case it does
-    for (column <- parquetColumn.children) {
-      checkColumn(column)
-    }
-
-    val sparkSchema = parquetColumn.sparkType.asInstanceOf[StructType]
-    parquetColumnVectors = (for (i <- 0 until sparkSchema.fields.length) yield {
-      new ParquetColumnVector(parquetColumn.children.apply(i),
-        vectors(i), capacity, MemoryMode.OFF_HEAP, missingColumns)
-    }).toArray
-
-    (totalRowCount, columnsRequested, cacheSchemaToReqSchemaMap, missingColumns,
-        columnsInCache, typesInCache)
-  }
+  val sparkSchema = parquetColumn.sparkType.asInstanceOf[StructType]
+  val parquetColumnVectors = (for (i <- 0 until sparkSchema.fields.length) yield {
+    new ParquetColumnVector(parquetColumn.children.apply(i),
+      vectors(i), capacity, MemoryMode.OFF_HEAP, missingColumns)
+  }).toArray
 
   private def containsPath(parquetType: Type, path: Array[String]): Boolean =
     containsPath(parquetType, path, 0)
@@ -265,28 +206,5 @@ class ShimCurrentBatchIterator(
     numBatched = num
     batchIdx = 0
     true
-  }
-
-  override def hasNext: Boolean = rowsReturned < totalRowCount
-
-  override def next(): ColumnarBatch = {
-    if (nextBatch) {
-      // FYI, A very IMPORTANT thing to note is that we are returning the columnar batch
-      // as-is i.e. this batch has NullTypes saved as IntegerTypes with null values. The
-      // way Spark optimizes the read of NullTypes makes this work without having to rip out
-      // the IntegerType column to be replaced by a NullType column. This could change in
-      // future and will affect this code.
-      columnarBatch
-    } else {
-      throw new NoSuchElementException("no elements found")
-    }
-  }
-
-  TaskContext.get().addTaskCompletionListener[Unit]((_: TaskContext) => {
-    close()
-  })
-
-  override def close(): Unit = {
-    parquetFileReader.close()
   }
 }
